+++ VpnHood.Client.App.UI.Html/src/i18n.js
@@ -86,7 +86,9 @@ const messages = {
         excludeLocalNetwork: "Exclude Local Network (LAN)",
         excludeLocalNetworkDesc: "Do not send my local network packets to VPN servers. I need to access the other computers and devices on my local network.",
         readPrivacyPolicy: "Read Privacy Policy",
-        privacyWarning: "IP logging is enabled on these servers. Please follow United States law, especially if using torrent."
+        privacyWarning: "These servers keep logs of IP activities for 30 days. To use these servers you must follow United States laws, especially when using torrent.",
+        createServer: "Create a Personal Server",
+        accept: "Accept"
     },
 };